requirements-dev.txt
-r requirements.txt
pytest
pytest-cov
requests-mock
isort
flake8
autopep8
tox
